+++ packages/block-library/src/image/index.js
@@ -106,7 +106,7 @@ if ( window.__experimentalContentOnlyInspectorFields ) {
 		},
 	];
 	settings[ formKey ] = {
-		fields: [ 'image' ],
+		fields: [ 'image', 'link', 'caption', 'alt' ],
 	};
 }