+++ packages/blocks/src/store/private-selectors.js
@@ -30,7 +30,6 @@ const ROOT_BLOCK_SUPPORTS = [
 	'blockGap',
 	'textAlign',
 	'textDecoration',
-	'textIndent',
 	'textTransform',
 	'letterSpacing',
 ];
@@ -84,11 +83,6 @@ function filterElementBlockSupports( blockSupports, name, element ) {
 				element === 'text'
 			)
 		) {
-			return false;
-		}
-
-		// Text indent is only available for blocks, not elements
-		if ( support === 'textIndent' && ! name ) {
 			return false;
 		}