+++ test/e2e/specs/editor/blocks/heading.spec.js
@@ -265,8 +265,10 @@ test.describe( 'Heading', () => {
 				name: 'core/heading',
 				attributes: {
 					content: 'Heading',
-					textAlign: 'center',
 					level: 4,
+					style: {
+						typography: { textAlign: 'center' },
+					},
 				},
 			},
 		] );
@@ -298,8 +300,10 @@ test.describe( 'Heading', () => {
 				name: 'core/heading',
 				attributes: {
 					content: 'Paragraph',
-					textAlign: 'center',
 					level: 2,
+					style: {
+						typography: { textAlign: 'center' },
+					},
 				},
 			},
 		] );
@@ -412,7 +416,7 @@ test.describe( 'Heading', () => {
 				);
 			} );
 
-			test( 'should preserve the text align attribute', async ( {
+			test( 'should preserve the text align block support', async ( {
 				editor,
 			} ) => {
 				await editor.insertBlock( {
@@ -425,7 +429,9 @@ test.describe( 'Heading', () => {
 				await editor.transformBlockTo( 'core/heading' );
 				const headingBlock = ( await editor.getBlocks() )[ 0 ];
 				expect( headingBlock.name ).toBe( 'core/heading' );
-				expect( headingBlock.attributes.textAlign ).toBe( 'right' );
+				expect(
+					headingBlock.attributes.style.typography.textAlign
+				).toBe( 'right' );
 			} );
 
 			test( 'should preserve the metadata attribute', async ( {
@@ -507,8 +513,8 @@ test.describe( 'Heading', () => {
 				await editor.insertBlock( {
 					name: 'core/heading',
 					attributes: {
-						textAlign: 'right',
 						content: 'initial content',
+						style: { typography: { textAlign: 'right' } },
 					},
 				} );
 				await editor.transformBlockTo( 'core/paragraph' );