+++ packages/ui/CHANGELOG.md
@@ -4,8 +4,13 @@
 
 ### Breaking Changes
 
--   Remove numeric multiplier option for spacing tokens from `Stack` and `Box` components. ([#73852](https://github.com/WordPress/gutenberg/pull/73852), [#74008](https://github.com/WordPress/gutenberg/pull/74008))
+-   Remove numeric multiplier option for spacing tokens from `Stack` and `Box` components ([#73852](https://github.com/WordPress/gutenberg/pull/73852), [#74008](https://github.com/WordPress/gutenberg/pull/74008)).
 
 ### New Features
 
--   Add `Stack` component. ([#73928](https://github.com/WordPress/gutenberg/pull/73928))
+-   Add `Stack` component ([#73928](https://github.com/WordPress/gutenberg/pull/73928)).
+-   Add `VisuallyHidden` component ([#74189](https://github.com/WordPress/gutenberg/pull/74189)).
+-   Add `Field` primitives ([#74190](https://github.com/WordPress/gutenberg/pull/74190)).
+-   Add `Fieldset` primitives ([#74296](https://github.com/WordPress/gutenberg/pull/74296)).
+-   Add `Icon` component ([#74311](https://github.com/WordPress/gutenberg/pull/74311)).
+-   Add `Button` component ([#74415](https://github.com/WordPress/gutenberg/pull/74415), [#74416](https://github.com/WordPress/gutenberg/pull/74416) ).